fix(win): open_file denied sharing, so a second read of an open container failed
_wfopen_s is not fopen: it opens the file with sharing denied, so while it
holds a handle nobody may open the file -- not even the same process. Every
read of an object inside a .pto goes through a second handle while PtoFile has
the container open, which is why build_cues, the ranged event read, the cue
rebuild, the generic EBML walk and the embedded-stream conformance case all
failed on Windows and nowhere else: not one of them could open the file. It is
_wfsopen with _SH_DENYNO now, which is what fopen does on every other platform.

Two more 32-bit value types while here: find_clsm_start_stop summed a macro
time into an unsigned long (32 bits on Windows, and a macro time passes 2^32
after a couple of minutes -- the function has no callers, so this is latent),
and the Coates pile-up correction truncated its photon count into a long.

The native-table range test asserted a 4x speed-up and Windows scored 3.85x:
the slice is dominated by the fixed cost of opening the container, which that
runner pays far more of. 2x still fails outright at the 1x a whole-object read
scores.

+++ modules/imaging/clsm/include/CLSMImage.h
@@ -65,7 +65,10 @@ static std::pair<int, int> find_clsm_start_stop(
             }
             else if(duration > 0) // search for stop idx using duration
             {
-                unsigned long stop_time = macro_time_arr[i_event] + duration;
+                // Not `unsigned long`: that is 32 bits on Windows, and a macro
+                // time passes 2^32 after a couple of minutes of acquisition.
+                unsigned long long stop_time =
+                        macro_time_arr[i_event] + (unsigned long long) duration;
                 for(;i_event < n; i_event++){
                     if(macro_time_arr[i_event] >= stop_time){
                         stop = i_event;
